+++ Assets/Scripts/Networking/Unity/UnityMessageEventDatabase/Database/UnityStringMessageEventDatabase.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Events;
 
-public class UnityStringMessageNetworkMessageEventsDatabase : UnityBaseMessageEventsDatabase<StringNetworkMessage, UnityStringMessageNetworkMessageEventsDatabase.StringCallbackWrapper, UnityStringMessageNetworkMessageEventsDatabase.StringMessageCallback>
+public class UnityStringMessageEventDatabase : UnityBaseMessageEventsDatabase<StringNetworkMessage, UnityStringMessageEventDatabase.StringCallbackWrapper, UnityStringMessageEventDatabase.StringMessageCallback>
 {
     [Serializable]
     public class StringCallbackWrapper : BaseMessageCallbackWrapper<StringNetworkMessage, StringMessageCallback>
