--- Assets/Scripts/Networking/Unity/UnityMessageEventDatabase/INetworkMessageEventsDatabase.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using GameFrame.Networking.Messaging.Message;
-using UnityEngine;
-using UnityEngine.Events;
-
-public interface INetworkMessageEventsDatabase<TMessage> where TMessage : NetworkMessage<NetworkEvent>
-{
-    Type GetDatabaseMessageType();
-
-    UnityEvent<TMessage> GetMessageCallback(NetworkEvent networkEvent);
-}
