--- Assets/Scripts/Networking/Unity/UnityMessageEventDatabase/INetworkMessageEventsDatabase.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e31b2965b3671a24d9d8c0479412fefe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
