--- Assets/Scripts/Networking/Unity/UnityMessageEventDatabase/UnityNetworkMessageEventCallbackDatabaseCollection.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 48ffabb67b93d744ea85e9475d24343f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
